Changing the cost data type

+++ GameShopLibrary/Item.cs
@@ -4,7 +4,7 @@ namespace GameShopLibrary
 {
     class Item
     {
-        public Item(int cost, Guid id, Game game)
+        public Item(double cost, Guid id, Game game)
         {
             Cost = cost;
             Id = id;
@@ -12,7 +12,7 @@ namespace GameShopLibrary
         }
 
         public Game Game { get; }
-        public int Cost { get; } 
+        public double Cost { get; } 
         public Guid Id { get; }
     }
